Add dialog title specification
The current implementation only allows default dialog titles like
"Open File" or "Select folder". In non-GUI applications that need to ask
for a number of files, it is convenient to state in the dialog title what
the purpose of the requested file is.
Unfortunately, this feature does depend on the submodule nativefiledialog.
I am not sure how to deal with that: a pull request is pending.

Signed-off-by: Len Remmerswaal <len.mail1@xs4all.nl>